fix ddl initialization and table access
