+++ pkgs/standards/autoapi/tests/i9n/test_allow_anon.py
@@ -66,7 +66,7 @@ def _build_client():
     api.set_auth(authn=auth.get_principal)
     auth.register_inject_hook(api)
     api.include_models([Tenant, Item])
-    api.initialize_sync()
+    api.initialize()
     app = App()
     app.include_router(api.router)
     prov = _resolver.resolve_provider()
@@ -98,7 +98,7 @@ def _build_client_attr():
     api.set_auth(authn=auth.get_principal)
     auth.register_inject_hook(api)
     api.include_models([Tenant, Item])
-    api.initialize_sync()
+    api.initialize()
     app = App()
     app.include_router(api.router)
     prov = _resolver.resolve_provider()
@@ -159,7 +159,7 @@ def _build_client_create_noauth():
     cfg = mem(async_=False)
     api = AutoApp(engine=cfg)
     api.include_models([Tenant, Item])
-    api.initialize_sync()
+    api.initialize()
     app = App()
     app.include_router(api.router)
     prov = _resolver.resolve_provider()
@@ -188,7 +188,7 @@ def _build_client_create_attr_noauth():
     cfg = mem(async_=False)
     api = AutoApp(engine=cfg)
     api.include_models([Tenant, Item])
-    api.initialize_sync()
+    api.initialize()
     app = App()
     app.include_router(api.router)
     prov = _resolver.resolve_provider()
