+++ pkgs/standards/autoapi/tests/i9n/test_v3_default_rpc_ops.py
@@ -44,7 +44,7 @@ async def client_and_model():
     api = AutoAPIv3(engine=mem())
     api.include_model(Gadget, prefix="")
     api.mount_jsonrpc(prefix="/rpc")
-    await api.initialize_async()
+    await api.initialize()
     app.include_router(api.router)
     transport = ASGITransport(app=app)
     client = AsyncClient(transport=transport, base_url="http://test")
@@ -174,7 +174,7 @@ async def bulk_client_and_model():
     api = AutoAPIv3(engine=mem())
     api.include_model(Gadget, prefix="")
     api.mount_jsonrpc(prefix="/rpc")
-    await api.initialize_async()
+    await api.initialize()
     app.include_router(api.router)
     transport = ASGITransport(app=app)
     client = AsyncClient(transport=transport, base_url="http://test")
