+++ pkgs/standards/autoapi/tests/unit/test_rpc_default_ops.py
@@ -28,7 +28,7 @@ def api_and_session() -> Iterator[tuple[AutoApp, Session, type[Base]]]:
 
     api = AutoApp(engine=mem(async_=False))
     api.include_model(Widget, mount_router=False)
-    api.initialize_sync()
+    api.initialize()
     prov = _resolver.resolve_provider()
     _, SessionLocal = prov.ensure()
