+++ pkgs/standards/autoapi/tests/unit/test_sqlite_attachments.py
@@ -10,7 +10,7 @@ def _db_names(conn):
 def test_initialize_sync_without_sqlite_attachments(sync_db_session):
     engine, get_db = sync_db_session
     api = AutoApp(get_db=get_db)
-    api.initialize_sync()
+    api.initialize()
     with engine.connect() as conn:
         assert _db_names(conn) == {"main"}
 
@@ -20,7 +20,7 @@ def test_initialize_sync_with_sqlite_attachments(sync_db_session, tmp_path):
     attach_db = tmp_path / "logs.sqlite"
     attach_db.touch()
     api = AutoApp(get_db=get_db)
-    api.initialize_sync(sqlite_attachments={"logs": str(attach_db)})
+    api.initialize(sqlite_attachments={"logs": str(attach_db)})
     with engine.connect() as conn:
         assert "logs" in _db_names(conn)
 
@@ -29,7 +29,7 @@ def test_initialize_sync_with_sqlite_attachments(sync_db_session, tmp_path):
 async def test_initialize_async_without_sqlite_attachments(async_db_session):
     engine, get_db = async_db_session
     api = AutoApp(get_db=get_db)
-    await api.initialize_async()
+    await api.initialize()
     async with engine.connect() as conn:
         result = await conn.exec_driver_sql("PRAGMA database_list")
         names = {row[1] for row in result.fetchall()}
@@ -42,7 +42,7 @@ async def test_initialize_async_with_sqlite_attachments(async_db_session, tmp_pa
     attach_db = tmp_path / "logs.sqlite"
     attach_db.touch()
     api = AutoApp(get_db=get_db)
-    await api.initialize_async(sqlite_attachments={"logs": str(attach_db)})
+    await api.initialize(sqlite_attachments={"logs": str(attach_db)})
     async with engine.connect() as conn:
         result = await conn.exec_driver_sql("PRAGMA database_list")
         names = {row[1] for row in result.fetchall()}
